+++ src/engine/LX_Vector2D.hpp
@@ -30,8 +30,8 @@
 */
 typedef struct LX_Vector2D
 {
-    int vx;                     /**< The x position */
-    int vy;                     /**< The y position */
+    float vx;                     /**< The x position */
+    float vy;                     /**< The y position */
 
 } LX_Vector2D;  /**< @brief The vector2D */
